requirements.txt
# LangGraph + LangChain for Multi-Agent Orchestration
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0

# Azure OpenAI Support
azure-identity

# State persistence and checkpointing
langgraph-checkpoint>=2.0.0
langgraph-checkpoint-sqlite>=2.0.0

# FastAPI for A2A server endpoints (optional)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# HTTP client
httpx>=0.27.0
aiohttp>=3.10.0

# GitHub API integration
PyGithub>=2.5.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Observability (optional - for LangSmith tracing)
langsmith>=0.1.0

# Async utilities
asyncio-atexit>=1.0.0

# Data processing
tiktoken>=0.8.0

# Vector embeddings (optional - uncomment based on needs)
# langchain-chroma>=0.1.0
# faiss-cpu>=1.8.0
